DDP-8465 | add a new method to ParticipantDao interface for retrieving participant_id by using the combination of guid and ddp_instance_id
